# [Enhancement] optimize ignore nulls