--- a/test/sql/test_window_function/R/test_ignore_nulls
+++ b/test/sql/test_window_function/R/test_ignore_nulls
@@ -1,0 +1,566 @@
+-- name: test_lead_lag_ignore_nulls
+CREATE TABLE `t0` (
+  `v1` int(11) NULL,
+  `v2` int(11) NULL,
+  `v3` int(11) NULL
+) ENGINE=OLAP
+DUPLICATE KEY(`v1`)
+DISTRIBUTED BY HASH(`v1`) BUCKETS 10
+PROPERTIES (
+  "replication_num" = "1"
+);
+-- result:
+-- !result
+INSERT INTO `t0` (v1, v2, v3) values
+    (1, 1, 1),
+    (1, 2, 2),
+    (1, 3, 3),
+    (1, 4, 4),
+    (2, 1, NULL),
+    (2, 2, 2),
+    (2, 3, 3),
+    (2, 4, 4),
+    (3, 1, 1),
+    (3, 2, NULL),
+    (3, 3, 3),
+    (3, 4, 4),
+    (4, 1, 1),
+    (4, 2, 2),
+    (4, 3, NULL),
+    (4, 4, 4),
+    (5, 1, 1),
+    (5, 2, 2),
+    (5, 3, 3),
+    (5, 4, NULL),
+    (6, 1, 1),
+    (6, 2, NULL),
+    (6, 3, NULL),
+    (6, 4, 4);
+-- result:
+-- !result
+SELECT v1, v2, v3, first_value(v3 IGNORE NULLS) OVER(ORDER BY v1, v2) FROM t0 ORDER BY v1, v2;
+-- result:
+1	1	1	1
+1	2	2	1
+1	3	3	1
+1	4	4	1
+2	1	None	1
+2	2	2	1
+2	3	3	1
+2	4	4	1
+3	1	1	1
+3	2	None	1
+3	3	3	1
+3	4	4	1
+4	1	1	1
+4	2	2	1
+4	3	None	1
+4	4	4	1
+5	1	1	1
+5	2	2	1
+5	3	3	1
+5	4	None	1
+6	1	1	1
+6	2	None	1
+6	3	None	1
+6	4	4	1
+-- !result
+SELECT v1, v2, v3, first_value(v3 IGNORE NULLS) OVER(ORDER BY v1, v2 rows between 1 preceding and 1 following) FROM t0 ORDER BY v1, v2;
+-- result:
+1	1	1	1
+1	2	2	1
+1	3	3	2
+1	4	4	3
+2	1	None	4
+2	2	2	2
+2	3	3	2
+2	4	4	3
+3	1	1	4
+3	2	None	1
+3	3	3	3
+3	4	4	3
+4	1	1	4
+4	2	2	1
+4	3	None	2
+4	4	4	4
+5	1	1	4
+5	2	2	1
+5	3	3	2
+5	4	None	3
+6	1	1	1
+6	2	None	1
+6	3	None	4
+6	4	4	4
+-- !result
+SELECT v1, v2, v3, first_value(v3 IGNORE NULLS) OVER(partition BY v1 ORDER BY v2) FROM t0 ORDER BY v1, v2;
+-- result:
+1	1	1	1
+1	2	2	1
+1	3	3	1
+1	4	4	1
+2	1	None	None
+2	2	2	2
+2	3	3	2
+2	4	4	2
+3	1	1	1
+3	2	None	1
+3	3	3	1
+3	4	4	1
+4	1	1	1
+4	2	2	1
+4	3	None	1
+4	4	4	1
+5	1	1	1
+5	2	2	1
+5	3	3	1
+5	4	None	1
+6	1	1	1
+6	2	None	1
+6	3	None	1
+6	4	4	1
+-- !result
+SELECT v1, v2, v3, first_value(v3 IGNORE NULLS) OVER(partition BY v1 ORDER BY v2 rows between 1 preceding and 1 following) FROM t0 ORDER BY v1, v2;
+-- result:
+1	1	1	1
+1	2	2	1
+1	3	3	2
+1	4	4	3
+2	1	None	2
+2	2	2	2
+2	3	3	2
+2	4	4	3
+3	1	1	1
+3	2	None	1
+3	3	3	3
+3	4	4	3
+4	1	1	1
+4	2	2	1
+4	3	None	2
+4	4	4	4
+5	1	1	1
+5	2	2	1
+5	3	3	2
+5	4	None	3
+6	1	1	1
+6	2	None	1
+6	3	None	4
+6	4	4	4
+-- !result
+SELECT v1, v2, v3, last_value(v3 IGNORE NULLS) OVER(ORDER BY v1, v2) FROM t0 ORDER BY v1, v2;
+-- result:
+1	1	1	1
+1	2	2	2
+1	3	3	3
+1	4	4	4
+2	1	None	4
+2	2	2	2
+2	3	3	3
+2	4	4	4
+3	1	1	1
+3	2	None	1
+3	3	3	3
+3	4	4	4
+4	1	1	1
+4	2	2	2
+4	3	None	2
+4	4	4	4
+5	1	1	1
+5	2	2	2
+5	3	3	3
+5	4	None	3
+6	1	1	1
+6	2	None	1
+6	3	None	1
+6	4	4	4
+-- !result
+SELECT v1, v2, v3, last_value(v3 IGNORE NULLS) OVER(ORDER BY v1, v2 rows between 1 preceding and 1 following) FROM t0 ORDER BY v1, v2;
+-- result:
+1	1	1	2
+1	2	2	3
+1	3	3	4
+1	4	4	4
+2	1	None	2
+2	2	2	3
+2	3	3	4
+2	4	4	1
+3	1	1	1
+3	2	None	3
+3	3	3	4
+3	4	4	1
+4	1	1	2
+4	2	2	2
+4	3	None	4
+4	4	4	1
+5	1	1	2
+5	2	2	3
+5	3	3	3
+5	4	None	1
+6	1	1	1
+6	2	None	1
+6	3	None	4
+6	4	4	4
+-- !result
+SELECT v1, v2, v3, last_value(v3 IGNORE NULLS) OVER(partition BY v1 ORDER BY v2) FROM t0 ORDER BY v1, v2;
+-- result:
+1	1	1	1
+1	2	2	2
+1	3	3	3
+1	4	4	4
+2	1	None	None
+2	2	2	2
+2	3	3	3
+2	4	4	4
+3	1	1	1
+3	2	None	1
+3	3	3	3
+3	4	4	4
+4	1	1	1
+4	2	2	2
+4	3	None	2
+4	4	4	4
+5	1	1	1
+5	2	2	2
+5	3	3	3
+5	4	None	3
+6	1	1	1
+6	2	None	1
+6	3	None	1
+6	4	4	4
+-- !result
+SELECT v1, v2, v3, last_value(v3 IGNORE NULLS) OVER(partition BY v1 ORDER BY v2 rows between 1 preceding and 1 following) FROM t0 ORDER BY v1, v2;
+-- result:
+1	1	1	2
+1	2	2	3
+1	3	3	4
+1	4	4	4
+2	1	None	2
+2	2	2	3
+2	3	3	4
+2	4	4	4
+3	1	1	1
+3	2	None	3
+3	3	3	4
+3	4	4	4
+4	1	1	2
+4	2	2	2
+4	3	None	4
+4	4	4	4
+5	1	1	2
+5	2	2	3
+5	3	3	3
+5	4	None	3
+6	1	1	1
+6	2	None	1
+6	3	None	4
+6	4	4	4
+-- !result
+SELECT v1, v2, v3, lead(v3 IGNORE NULLS, 1) OVER(ORDER BY v1, v2) FROM t0 ORDER BY v1, v2;
+-- result:
+1	1	1	2
+1	2	2	3
+1	3	3	4
+1	4	4	2
+2	1	None	2
+2	2	2	3
+2	3	3	4
+2	4	4	1
+3	1	1	3
+3	2	None	3
+3	3	3	4
+3	4	4	1
+4	1	1	2
+4	2	2	4
+4	3	None	4
+4	4	4	1
+5	1	1	2
+5	2	2	3
+5	3	3	1
+5	4	None	1
+6	1	1	4
+6	2	None	4
+6	3	None	4
+6	4	4	None
+-- !result
+SELECT v1, v2, v3, lead(v3 IGNORE NULLS, 1) OVER(partition BY v1 ORDER BY v2) FROM t0 ORDER BY v1, v2;
+-- result:
+1	1	1	2
+1	2	2	3
+1	3	3	4
+1	4	4	None
+2	1	None	2
+2	2	2	3
+2	3	3	4
+2	4	4	None
+3	1	1	3
+3	2	None	3
+3	3	3	4
+3	4	4	None
+4	1	1	2
+4	2	2	4
+4	3	None	4
+4	4	4	None
+5	1	1	2
+5	2	2	3
+5	3	3	None
+5	4	None	None
+6	1	1	4
+6	2	None	4
+6	3	None	4
+6	4	4	None
+-- !result
+SELECT v1, v2, v3, lead(v3 IGNORE NULLS, 2) OVER(ORDER BY v1, v2) FROM t0 ORDER BY v1, v2;
+-- result:
+1	1	1	3
+1	2	2	4
+1	3	3	2
+1	4	4	3
+2	1	None	3
+2	2	2	4
+2	3	3	1
+2	4	4	3
+3	1	1	4
+3	2	None	4
+3	3	3	1
+3	4	4	2
+4	1	1	4
+4	2	2	1
+4	3	None	1
+4	4	4	2
+5	1	1	3
+5	2	2	1
+5	3	3	4
+5	4	None	4
+6	1	1	None
+6	2	None	None
+6	3	None	None
+6	4	4	None
+-- !result
+SELECT v1, v2, v3, lead(v3 IGNORE NULLS, 2) OVER(partition BY v1 ORDER BY v2) FROM t0 ORDER BY v1, v2;
+-- result:
+1	1	1	3
+1	2	2	4
+1	3	3	None
+1	4	4	None
+2	1	None	3
+2	2	2	4
+2	3	3	None
+2	4	4	None
+3	1	1	4
+3	2	None	4
+3	3	3	None
+3	4	4	None
+4	1	1	4
+4	2	2	None
+4	3	None	None
+4	4	4	None
+5	1	1	3
+5	2	2	None
+5	3	3	None
+5	4	None	None
+6	1	1	None
+6	2	None	None
+6	3	None	None
+6	4	4	None
+-- !result
+SELECT v1, v2, v3, lag(v3 IGNORE NULLS, 1) OVER(ORDER BY v1, v2) FROM t0 ORDER BY v1, v2;
+-- result:
+1	1	1	None
+1	2	2	1
+1	3	3	2
+1	4	4	3
+2	1	None	4
+2	2	2	4
+2	3	3	2
+2	4	4	3
+3	1	1	4
+3	2	None	1
+3	3	3	1
+3	4	4	3
+4	1	1	4
+4	2	2	1
+4	3	None	2
+4	4	4	2
+5	1	1	4
+5	2	2	1
+5	3	3	2
+5	4	None	3
+6	1	1	3
+6	2	None	1
+6	3	None	1
+6	4	4	1
+-- !result
+SELECT v1, v2, v3, lag(v3 IGNORE NULLS, 1) OVER(partition BY v1 ORDER BY v2) FROM t0 ORDER BY v1, v2;
+-- result:
+1	1	1	None
+1	2	2	1
+1	3	3	2
+1	4	4	3
+2	1	None	None
+2	2	2	None
+2	3	3	2
+2	4	4	3
+3	1	1	None
+3	2	None	1
+3	3	3	1
+3	4	4	3
+4	1	1	None
+4	2	2	1
+4	3	None	2
+4	4	4	2
+5	1	1	None
+5	2	2	1
+5	3	3	2
+5	4	None	3
+6	1	1	None
+6	2	None	1
+6	3	None	1
+6	4	4	1
+-- !result
+SELECT v1, v2, v3, lag(v3 IGNORE NULLS, 2) OVER(ORDER BY v1, v2) FROM t0 ORDER BY v1, v2;
+-- result:
+1	1	1	None
+1	2	2	None
+1	3	3	1
+1	4	4	2
+2	1	None	3
+2	2	2	3
+2	3	3	4
+2	4	4	2
+3	1	1	3
+3	2	None	4
+3	3	3	4
+3	4	4	1
+4	1	1	3
+4	2	2	4
+4	3	None	1
+4	4	4	1
+5	1	1	2
+5	2	2	4
+5	3	3	1
+5	4	None	2
+6	1	1	2
+6	2	None	3
+6	3	None	3
+6	4	4	3
+-- !result
+SELECT v1, v2, v3, lag(v3 IGNORE NULLS, 2) OVER(partition BY v1 ORDER BY v2) FROM t0 ORDER BY v1, v2;
+-- result:
+1	1	1	None
+1	2	2	None
+1	3	3	1
+1	4	4	2
+2	1	None	None
+2	2	2	None
+2	3	3	None
+2	4	4	2
+3	1	1	None
+3	2	None	None
+3	3	3	None
+3	4	4	1
+4	1	1	None
+4	2	2	None
+4	3	None	1
+4	4	4	1
+5	1	1	None
+5	2	2	None
+5	3	3	1
+5	4	None	2
+6	1	1	None
+6	2	None	None
+6	3	None	None
+6	4	4	None
+-- !result
+-- name: test_lead_lag_ignore_nulls_all_nulls
+CREATE TABLE `t_all_null` (
+  `v1` int(11) NULL,
+  `v2` int(11) NULL,
+  `v3` int(11) NULL
+) ENGINE=OLAP
+DUPLICATE KEY(`v1`)
+DISTRIBUTED BY HASH(`v1`) BUCKETS 10
+PROPERTIES (
+  "replication_num" = "1"
+);
+-- result:
+-- !result
+INSERT INTO `t_all_null` (v1, v2, v3) values
+    (1, 1, NULL),
+    (1, 2, NULL),
+    (1, 3, NULL),
+    (1, 4, NULL),
+    (2, 1, NULL),
+    (2, 2, NULL),
+    (2, 3, NULL),
+    (2, 4, NULL),
+    (3, 1, NULL),
+    (3, 2, NULL),
+    (3, 3, NULL),
+    (3, 4, NULL),
+    (4, 1, NULL),
+    (4, 2, NULL),
+    (4, 3, NULL),
+    (4, 4, NULL),
+    (5, 1, NULL),
+    (5, 2, NULL),
+    (5, 3, NULL),
+    (5, 4, NULL),
+    (6, 1, NULL),
+    (6, 2, NULL),
+    (6, 3, NULL),
+    (6, 4, NULL);
+-- result:
+-- !result
+SELECT v1, v2, v3, lead(v3 IGNORE NULLS, 1) OVER(partition BY v1 ORDER BY v2) FROM t_all_null ORDER BY v1, v2;
+-- result:
+1	1	None	None
+1	2	None	None
+1	3	None	None
+1	4	None	None
+2	1	None	None
+2	2	None	None
+2	3	None	None
+2	4	None	None
+3	1	None	None
+3	2	None	None
+3	3	None	None
+3	4	None	None
+4	1	None	None
+4	2	None	None
+4	3	None	None
+4	4	None	None
+5	1	None	None
+5	2	None	None
+5	3	None	None
+5	4	None	None
+6	1	None	None
+6	2	None	None
+6	3	None	None
+6	4	None	None
+-- !result
+SELECT v1, v2, v3, lag(v3 IGNORE NULLS, 1) OVER(partition BY v1 ORDER BY v2) FROM t_all_null ORDER BY v1, v2;
+-- result:
+1	1	None	None
+1	2	None	None
+1	3	None	None
+1	4	None	None
+2	1	None	None
+2	2	None	None
+2	3	None	None
+2	4	None	None
+3	1	None	None
+3	2	None	None
+3	3	None	None
+3	4	None	None
+4	1	None	None
+4	2	None	None
+4	3	None	None
+4	4	None	None
+5	1	None	None
+5	2	None	None
+5	3	None	None
+5	4	None	None
+6	1	None	None
+6	2	None	None
+6	3	None	None
+6	4	None	None
+-- !result