## Why I'm doing:
lead/lag ignore nulls' implementation is O(N^2) at worst time when partition is all null. 
For example, every round lag will try to find 'offset' non-null elments before current row, and found nothing when visit the first element in partition.

## What I'm doing:
use two fileds to remeber some status before this round, so time complexity is O(N)

this is my test result, one thread with one partition, whose row number is increseing, you can see the different

```
row number| before | after
--     |  --      |     --
10w | 2.34s | 0.33s
20w | 9.20s | 0.60s
40w | 36.40s | 1.16s
```

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
- [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [x] 3.3
  - [x] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5